requirements.txt
Flask==2.3.3
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
mysql-connector-python==8.1.0
redis==5.0.1
cryptography==41.0.7
PyJWT==2.8.0
Werkzeug==2.3.7
gunicorn==21.2.0
PyMySQL==1.1.1
xmltodict==0.14.2
xmltodict==0.14.2